// src/types/Transaction.ts
export type TransactionType = "Payment" | "Transfer" | "Credit" | "Withdrawal";
export type TransactionStatus = "Pending" | "Success" | "Failed";

export interface Transaction {
  id: string;
  type: TransactionType;
  title: string;
  amount: number;
  currency: string;
  transactionName: string;
  transactionDescription: string;
  date: string;
  status: TransactionStatus;
  cardNumberUsed: string | null;
  authorizedUser: string | null;
  icon: string;
  color: string;
}